+++ js/llmr.js
@@ -16,6 +16,7 @@ if (typeof window === 'undefined') {
     llmr.StyleDeclaration = require('./style/styledeclaration.js');
 
     llmr.LatLng = require('./geometry/latlng.js');
+    llmr.LatLngBounds = require('./geometry/latlngbounds.js');
     llmr.Point = require('./geometry/point.js');
 
     llmr.Tile = require('./ui/tile.js');