+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - **breaking**: llmr now uses a completely **new style format**.
 To migrate old styles, use the [gl-style script](https://github.com/mapbox/gl-style)
+- `map.fitBounds` now accepts `bounds` (`[[minLat, minLng], [maxLat, maxLng]]`) instead of `minLat, minLng, maxLat, maxLng` as arguments.
+- Added `LatLngBounds` geometry type.
 - **breaking**: round linejoins are now specified with `"join": "round"` on the bucket,
   and they no longer need `"linejoin": "round"` in the style.
 - Zoom value used in styles now gets adjusted based on latitude. Disabled by `adjustZoom: false` in map options.